ARTEMIS-2984 Simplify large message controllers (re)moving dead code
