+++ artemis-core-client/src/main/java/org/apache/activemq/artemis/core/client/impl/LargeMessageControllerImpl.java
@@ -5,9 +5,9 @@
  * The ASF licenses this file to You under the Apache License, Version 2.0
  * (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <br>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <br>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,331 +17,39 @@
 package org.apache.activemq.artemis.core.client.impl;
 
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
 import java.nio.ByteBuffer;
-import java.nio.channels.FileChannel;
-import java.nio.channels.GatheringByteChannel;
-import java.nio.channels.ScatteringByteChannel;
-import java.nio.file.StandardOpenOption;
-import java.util.concurrent.LinkedBlockingQueue;
-import java.util.concurrent.TimeUnit;
 
 import io.netty.buffer.ByteBuf;
 import org.apache.activemq.artemis.api.core.ActiveMQBuffer;
-import org.apache.activemq.artemis.api.core.ActiveMQException;
-import org.apache.activemq.artemis.api.core.ActiveMQExceptionType;
-import org.apache.activemq.artemis.api.core.ActiveMQInterruptedException;
-import org.apache.activemq.artemis.api.core.ActiveMQLargeMessageInterruptedException;
 import org.apache.activemq.artemis.api.core.SimpleString;
 import org.apache.activemq.artemis.core.client.ActiveMQClientLogger;
-import org.apache.activemq.artemis.core.client.ActiveMQClientMessageBundle;
 import org.apache.activemq.artemis.utils.ByteUtil;
 import org.apache.activemq.artemis.utils.DataConstants;
 import org.apache.activemq.artemis.utils.UTF8Util;
 
-/**
- * This class aggregates several {@link org.apache.activemq.artemis.core.protocol.core.impl.wireformat.SessionReceiveContinuationMessage}
- * as it was being handled
- * by a single buffer. This buffer can be consumed as messages are arriving, and it will hold the
- * packets until they are read using the ChannelBuffer interface, or the setOutputStream or
- * saveStream are called.
- */
-public class LargeMessageControllerImpl implements LargeMessageController {
-   // Constants -----------------------------------------------------
+public final class LargeMessageControllerImpl extends LargeMessageControllerBaseImpl {
 
    private static final String READ_ONLY_ERROR_MESSAGE = "This is a read-only buffer, setOperations are not supported";
-
-   // Attributes ----------------------------------------------------
-
-   private final ClientConsumerInternal consumerInternal;
-
-   private final LinkedBlockingQueue<LargeData> largeMessageData = new LinkedBlockingQueue<>();
-
-   private volatile LargeData currentPacket = null;
-
-   private final long totalSize;
-
-   private final int bufferSize;
-
-   private boolean streamEnded = false;
-
-   private boolean streamClosed = false;
-
-   private final long readTimeout;
-
    private long readerIndex = 0;
 
-   /**
-    * This is to control if packets are arriving for a better timeout control
-    */
-   private boolean packetAdded = false;
-
-   private long packetPosition = -1;
-
-   private long lastIndex = 0;
-
-   private long packetLastPosition = -1;
-
-   private OutputStream outStream;
-
-   // There's no need to wait a synchronization
-   // we just set the exception and let other threads to get it as soon as possible
-   private volatile Exception handledException;
-
-   private final FileCache fileCache;
-
-   private boolean local = false;
-   // Static --------------------------------------------------------
-
-   // Constructors --------------------------------------------------
-
-   public LargeMessageControllerImpl(final ClientConsumerInternal consumerInternal,
-                                     final long totalSize,
-                                     final long readTimeout) {
-      this(consumerInternal, totalSize, readTimeout, null);
+   public LargeMessageControllerImpl(ClientConsumerInternal consumerInternal, long totalSize, long readTimeout) {
+      super(consumerInternal, totalSize, readTimeout);
    }
 
-   public LargeMessageControllerImpl(final ClientConsumerInternal consumerInternal,
-                                     final long totalSize,
-                                     final long readTimeout,
-                                     final File cachedFile) {
-      this(consumerInternal, totalSize, readTimeout, cachedFile, 10 * 1024);
+   public LargeMessageControllerImpl(ClientConsumerInternal consumerInternal,
+                                     long totalSize,
+                                     long readTimeout,
+                                     File cachedFile) {
+      super(consumerInternal, totalSize, readTimeout, cachedFile);
    }
 
-   public LargeMessageControllerImpl(final ClientConsumerInternal consumerInternal,
-                                     final long totalSize,
-                                     final long readTimeout,
-                                     final File cachedFile,
-                                     final int bufferSize) {
-      this.consumerInternal = consumerInternal;
-      this.readTimeout = readTimeout;
-      this.totalSize = totalSize;
-      if (cachedFile == null) {
-         fileCache = null;
-      } else {
-         fileCache = new FileCache(cachedFile);
-      }
-      this.bufferSize = bufferSize;
-   }
-
-   // Public --------------------------------------------------------
-
-   public void setLocal(boolean local) {
-      this.local = local;
-   }
-
-   @Override
-   public void discardUnusedPackets() {
-      if (outStream == null) {
-         if (local)
-            return;
-         try {
-            checkForPacket(totalSize - 1);
-         } catch (Throwable ignored) {
-         }
-      }
-   }
-
-   /**
-    * TODO: move this to ConsumerContext as large message is a protocol specific thing
-    * Add a buff to the List, or save it to the OutputStream if set
-    */
-   @Override
-   public void addPacket(byte[] chunk, int flowControlSize, boolean isContinues) {
-      int flowControlCredit = 0;
-
-      synchronized (this) {
-         packetAdded = true;
-         if (outStream != null) {
-            try {
-               if (!isContinues) {
-                  streamEnded = true;
-               }
-
-               if (fileCache != null) {
-                  fileCache.cachePackage(chunk);
-               }
-
-               outStream.write(chunk);
-
-               flowControlCredit = flowControlSize;
-
-               notifyAll();
-
-               if (streamEnded) {
-                  outStream.close();
-               }
-            } catch (Exception e) {
-               ActiveMQClientLogger.LOGGER.errorAddingPacket(e);
-               handledException = e;
-            }
-         } else {
-            if (fileCache != null) {
-               try {
-                  fileCache.cachePackage(chunk);
-               } catch (Exception e) {
-                  ActiveMQClientLogger.LOGGER.errorAddingPacket(e);
-                  handledException = e;
-               }
-            }
-
-            largeMessageData.offer(new LargeData(chunk, flowControlSize, isContinues));
-         }
-      }
-
-      if (flowControlCredit != 0) {
-         try {
-            consumerInternal.flowControl(flowControlCredit, !isContinues);
-         } catch (Exception e) {
-            ActiveMQClientLogger.LOGGER.errorAddingPacket(e);
-            handledException = e;
-         }
-      }
-   }
-
-   @Override
-   public void cancel() {
-      this.handledException = ActiveMQClientMessageBundle.BUNDLE.largeMessageInterrupted();
-
-      synchronized (this) {
-         int totalSize = 0;
-         LargeData polledPacket = null;
-         while ((polledPacket = largeMessageData.poll()) != null) {
-            totalSize += polledPacket.getFlowControlSize();
-         }
-
-         try {
-            consumerInternal.flowControl(totalSize, false);
-         } catch (Exception ignored) {
-            // what else can we do here?
-            ActiveMQClientLogger.LOGGER.errorCallingCancel(ignored);
-         }
-
-         largeMessageData.offer(new LargeData());
-         streamEnded = true;
-         streamClosed = true;
-
-         notifyAll();
-      }
-   }
-
-   @Override
-   public synchronized void close() {
-      if (fileCache != null) {
-         fileCache.close();
-      }
-   }
-
-   @Override
-   public void setOutputStream(final OutputStream output) throws ActiveMQException {
-
-      int totalFlowControl = 0;
-      boolean continues = false;
-
-      synchronized (this) {
-         if (currentPacket != null) {
-            sendPacketToOutput(output, currentPacket);
-            currentPacket = null;
-         }
-         while (handledException == null) {
-            LargeData packet = largeMessageData.poll();
-            if (packet == null) {
-               break;
-            }
-            totalFlowControl += packet.getFlowControlSize();
-
-            continues = packet.isContinues();
-            sendPacketToOutput(output, packet);
-         }
-
-         checkException();
-         outStream = output;
-      }
-
-      if (totalFlowControl > 0) {
-         consumerInternal.flowControl(totalFlowControl, !continues);
-      }
-   }
-
-   @Override
-   public synchronized void saveBuffer(final OutputStream output) throws ActiveMQException {
-      if (streamClosed) {
-         throw ActiveMQClientMessageBundle.BUNDLE.largeMessageLostSession();
-      }
-      setOutputStream(output);
-      waitCompletion(0);
-   }
-
-   /**
-    * @param timeWait Milliseconds to Wait. 0 means forever
-    * @throws ActiveMQException
-    */
-   @Override
-   public synchronized boolean waitCompletion(final long timeWait) throws ActiveMQException {
-      if (outStream == null) {
-         // There is no stream.. it will never achieve the end of streaming
-         return false;
-      }
-
-      long timeOut;
-
-      // If timeWait = 0, we will use the readTimeout
-      // And we will check if no packets have arrived within readTimeout milliseconds
-      if (timeWait != 0) {
-         timeOut = System.currentTimeMillis() + timeWait;
-      } else {
-         timeOut = System.currentTimeMillis() + readTimeout;
-      }
-
-      while (!streamEnded && handledException == null) {
-         try {
-            this.wait(timeWait == 0 ? readTimeout : timeWait);
-         } catch (InterruptedException e) {
-            throw new ActiveMQInterruptedException(e);
-         }
-
-         if (!streamEnded && handledException == null) {
-            if (timeWait != 0 && System.currentTimeMillis() > timeOut) {
-               throw ActiveMQClientMessageBundle.BUNDLE.timeoutOnLargeMessage();
-            } else if (System.currentTimeMillis() > timeOut && !packetAdded) {
-               throw ActiveMQClientMessageBundle.BUNDLE.timeoutOnLargeMessage();
-            }
-         }
-      }
-
-      checkException();
-
-      return streamEnded;
-
-   }
-
-   /**
-    * @throws ActiveMQException
-    */
-   private void checkException() throws ActiveMQException {
-      // it's not needed to copy it as we never set it back to null
-      // once the exception is set, the controller is pretty much useless
-      if (handledException != null) {
-         if (handledException instanceof ActiveMQException) {
-            ActiveMQException nestedException;
-
-            // This is just to be user friendly and give the user a proper exception trace,
-            // instead to just where it was canceled.
-            if (handledException instanceof ActiveMQLargeMessageInterruptedException) {
-               nestedException = new ActiveMQLargeMessageInterruptedException(handledException.getMessage());
-            } else {
-               nestedException = new ActiveMQException(((ActiveMQException) handledException).getType(), handledException.getMessage());
-            }
-            nestedException.initCause(handledException);
-
-            throw nestedException;
-         } else {
-            throw new ActiveMQException(ActiveMQExceptionType.LARGE_MESSAGE_ERROR_BODY, "Error on saving LargeMessageBufferImpl", handledException);
-         }
-      }
+   public LargeMessageControllerImpl(ClientConsumerInternal consumerInternal,
+                                     long totalSize,
+                                     long readTimeout,
+                                     File cachedFile,
+                                     int bufferSize) {
+      super(consumerInternal, totalSize, readTimeout, cachedFile, bufferSize);
    }
 
    // Channel Buffer Implementation ---------------------------------
@@ -361,21 +69,9 @@ public class LargeMessageControllerImpl implements LargeMessageController {
       return getByte((long) index);
    }
 
-   private byte getByte(final long index) {
-      checkForPacket(index);
-
-      if (fileCache != null && index < packetPosition) {
-         return fileCache.getByteFromCache(index);
-      } else {
-         return currentPacket.getChunk()[(int) (index - packetPosition)];
-      }
-   }
-
    @Override
    public void getBytes(final int index, final ActiveMQBuffer dst, final int dstIndex, final int length) {
-      byte[] destBytes = new byte[length];
-      getBytes(index, destBytes);
-      dst.setBytes(dstIndex, destBytes);
+      getBytes((long) index, dst, dstIndex, length);
    }
 
    private void getBytes(final long index, final ActiveMQBuffer dst, final int dstIndex, final int length) {
@@ -386,14 +82,10 @@ public class LargeMessageControllerImpl implements LargeMessageController {
 
    @Override
    public void getBytes(final int index, final byte[] dst, final int dstIndex, final int length) {
-      byte[] bytesToGet = new byte[length];
-
-      getBytes(index, bytesToGet);
-
-      System.arraycopy(bytesToGet, 0, dst, dstIndex, length);
+      getBytes((long) index, dst, dstIndex, length);
    }
 
-   public void getBytes(final long index, final byte[] dst, final int dstIndex, final int length) {
+   private void getBytes(final long index, final byte[] dst, final int dstIndex, final int length) {
       byte[] bytesToGet = new byte[length];
 
       getBytes(index, bytesToGet);
@@ -403,43 +95,21 @@ public class LargeMessageControllerImpl implements LargeMessageController {
 
    @Override
    public void getBytes(final int index, final ByteBuffer dst) {
+      getBytes((long) index, dst);
+   }
+
+   private void getBytes(final long index, final ByteBuffer dst) {
       byte[] bytesToGet = new byte[dst.remaining()];
       getBytes(index, bytesToGet);
       dst.put(bytesToGet);
-   }
-
-   public void getBytes(final long index, final ByteBuffer dst) {
-      byte[] bytesToGet = new byte[dst.remaining()];
-      getBytes(index, bytesToGet);
-      dst.put(bytesToGet);
-   }
-
-   public void getBytes(final int index, final OutputStream out, final int length) throws IOException {
-      byte[] bytesToGet = new byte[length];
-      getBytes(index, bytesToGet);
-      out.write(bytesToGet);
-   }
-
-   public void getBytes(final long index, final OutputStream out, final int length) throws IOException {
-      byte[] bytesToGet = new byte[length];
-      getBytes(index, bytesToGet);
-      out.write(bytesToGet);
-   }
-
-   public int getBytes(final int index, final GatheringByteChannel out, final int length) throws IOException {
-      byte[] bytesToGet = new byte[length];
-      getBytes(index, bytesToGet);
-      return out.write(ByteBuffer.wrap(bytesToGet));
    }
 
    @Override
    public int getInt(final int index) {
-      return (getByte(index) & 0xff) << 24 | (getByte(index + 1) & 0xff) << 16 |
-         (getByte(index + 2) & 0xff) << 8 |
-         (getByte(index + 3) & 0xff) << 0;
+      return getInt((long) index);
    }
 
-   public int getInt(final long index) {
+   private int getInt(final long index) {
       return (getByte(index) & 0xff) << 24 | (getByte(index + 1) & 0xff) << 16 |
          (getByte(index + 2) & 0xff) << 8 |
          (getByte(index + 3) & 0xff) << 0;
@@ -447,16 +117,10 @@ public class LargeMessageControllerImpl implements LargeMessageController {
 
    @Override
    public long getLong(final int index) {
-      return ((long) getByte(index) & 0xff) << 56 | ((long) getByte(index + 1) & 0xff) << 48 |
-         ((long) getByte(index + 2) & 0xff) << 40 |
-         ((long) getByte(index + 3) & 0xff) << 32 |
-         ((long) getByte(index + 4) & 0xff) << 24 |
-         ((long) getByte(index + 5) & 0xff) << 16 |
-         ((long) getByte(index + 6) & 0xff) << 8 |
-         ((long) getByte(index + 7) & 0xff) << 0;
+      return getLong((long) index);
    }
 
-   public long getLong(final long index) {
+   private long getLong(final long index) {
       return ((long) getByte(index) & 0xff) << 56 | ((long) getByte(index + 1) & 0xff) << 48 |
          ((long) getByte(index + 2) & 0xff) << 40 |
          ((long) getByte(index + 3) & 0xff) << 32 |
@@ -468,59 +132,51 @@ public class LargeMessageControllerImpl implements LargeMessageController {
 
    @Override
    public short getShort(final int index) {
+      return getShort((long) index);
+   }
+
+   private short getShort(final long index) {
       return (short) (getByte(index) << 8 | getByte(index + 1) & 0xFF);
-   }
-
-   public short getShort(final long index) {
-      return (short) (getByte(index) << 8 | getByte(index + 1) & 0xFF);
-   }
-
-   private int getUnsignedMedium(final int index) {
-      return (getByte(index) & 0xff) << 16 | (getByte(index + 1) & 0xff) << 8 | (getByte(index + 2) & 0xff) << 0;
-   }
-
-   public int getUnsignedMedium(final long index) {
-      return (getByte(index) & 0xff) << 16 | (getByte(index + 1) & 0xff) << 8 | (getByte(index + 2) & 0xff) << 0;
    }
 
    @Override
    public void setByte(final int index, final byte value) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void setBytes(final int index, final ActiveMQBuffer src, final int srcIndex, final int length) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void setBytes(final int index, final byte[] src, final int srcIndex, final int length) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void setBytes(final int index, final ByteBuffer src) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void setInt(final int index, final int value) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void setLong(final int index, final long value) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void setShort(final int index, final short value) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public ByteBuffer toByteBuffer(final int index, final int length) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
@@ -546,17 +202,12 @@ public class LargeMessageControllerImpl implements LargeMessageController {
 
    @Override
    public int writerIndex() {
-      return (int) totalSize;
-   }
-
-   @Override
-   public long getSize() {
-      return totalSize;
+      return (int) getSize();
    }
 
    @Override
    public void writerIndex(final int writerIndex) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
@@ -586,12 +237,12 @@ public class LargeMessageControllerImpl implements LargeMessageController {
 
    @Override
    public int readableBytes() {
-      long readableBytes = totalSize - readerIndex;
+      long readableBytes = getSize() - readerIndex;
 
       if (readableBytes > Integer.MAX_VALUE) {
          return Integer.MAX_VALUE;
       } else {
-         return (int) (totalSize - readerIndex);
+         return (int) (getSize() - readerIndex);
       }
    }
 
@@ -602,7 +253,7 @@ public class LargeMessageControllerImpl implements LargeMessageController {
 
    @Override
    public void markReaderIndex() {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
@@ -617,17 +268,17 @@ public class LargeMessageControllerImpl implements LargeMessageController {
 
    @Override
    public void markWriterIndex() {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void resetWriterIndex() {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void discardReadBytes() {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
@@ -640,14 +291,6 @@ public class LargeMessageControllerImpl implements LargeMessageController {
       return getShort(index) & 0xFFFF;
    }
 
-   public int getMedium(final int index) {
-      int value = getUnsignedMedium(index);
-      if ((value & 0x800000) != 0) {
-         value |= 0xff000000;
-      }
-      return value;
-   }
-
    @Override
    public long getUnsignedInt(final int index) {
       return getInt(index) & 0xFFFFFFFFL;
@@ -655,13 +298,10 @@ public class LargeMessageControllerImpl implements LargeMessageController {
 
    @Override
    public void getBytes(int index, final byte[] dst) {
-      // TODO: optimize this by using System.arraycopy
-      for (int i = 0; i < dst.length; i++) {
-         dst[i] = getByte(index++);
-      }
+      getBytes((long) index, dst);
    }
 
-   public void getBytes(long index, final byte[] dst) {
+   private void getBytes(long index, final byte[] dst) {
       // TODO: optimize this by using System.arraycopy
       for (int i = 0; i < dst.length; i++) {
          dst[i] = getByte(index++);
@@ -684,21 +324,17 @@ public class LargeMessageControllerImpl implements LargeMessageController {
 
    @Override
    public void setBytes(final int index, final byte[] src) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void setBytes(final int index, final ActiveMQBuffer src) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void setBytes(final int index, final ActiveMQBuffer src, final int length) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
-   }
-
-   public void setZero(final int index, final int length) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
@@ -718,29 +354,11 @@ public class LargeMessageControllerImpl implements LargeMessageController {
       return readShort() & 0xFFFF;
    }
 
-   public int readMedium() {
-      int value = readUnsignedMedium();
-      if ((value & 0x800000) != 0) {
-         value |= 0xff000000;
-      }
-      return value;
-   }
-
-   public int readUnsignedMedium() {
-      int v = getUnsignedMedium(readerIndex);
-      readerIndex += 3;
-      return v;
-   }
-
    @Override
    public int readInt() {
       int v = getInt(readerIndex);
       readerIndex += 4;
       return v;
-   }
-
-   public int readInt(final int pos) {
-      return getInt(pos);
    }
 
    @Override
@@ -813,17 +431,6 @@ public class LargeMessageControllerImpl implements LargeMessageController {
       readerIndex += length;
    }
 
-   public int readBytes(final GatheringByteChannel out, final int length) throws IOException {
-      int readBytes = getBytes((int) readerIndex, out, length);
-      readerIndex += readBytes;
-      return readBytes;
-   }
-
-   public void readBytes(final OutputStream out, final int length) throws IOException {
-      getBytes(readerIndex, out, length);
-      readerIndex += length;
-   }
-
    @Override
    public int skipBytes(final int length) {
 
@@ -835,60 +442,52 @@ public class LargeMessageControllerImpl implements LargeMessageController {
 
    @Override
    public void writeByte(final byte value) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void writeShort(final short value) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
-   }
-
-   public void writeMedium(final int value) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void writeInt(final int value) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void writeNullableInt(final Integer value) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void writeLong(final long value) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void writeNullableLong(final Long value) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void writeBytes(final byte[] src, final int srcIndex, final int length) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void writeBytes(final byte[] src) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
-   }
-
-   public void writeBytes(final ActiveMQBuffer src) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void writeBytes(final ActiveMQBuffer src, final int length) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void writeBytes(final ByteBuffer src) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    /**
@@ -903,40 +502,12 @@ public class LargeMessageControllerImpl implements LargeMessageController {
     */
    @Override
    public void writeBytes(ByteBuf src, int srcIndex, int length) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
-   }
-
-   public int writeBytes(final InputStream in, final int length) throws IOException {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
-   }
-
-   public int writeBytes(final ScatteringByteChannel in, final int length) throws IOException {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
-   }
-
-   public void writeZero(final int length) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public ByteBuffer toByteBuffer() {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
-   }
-
-   public ByteBuffer[] toByteBuffers() {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
-   }
-
-   public ByteBuffer[] toByteBuffers(final int index, final int length) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
-   }
-
-   public String toString(final String charsetName) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
-   }
-
-   public Object getUnderlyingBuffer() {
-      return this;
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
@@ -1036,54 +607,54 @@ public class LargeMessageControllerImpl implements LargeMessageController {
 
    @Override
    public void writeBoolean(final boolean val) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void writeNullableBoolean(final Boolean val) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void writeChar(final char val) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void writeDouble(final double val) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
 
    }
 
    @Override
    public void writeFloat(final float val) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
 
    }
 
    @Override
    public void writeNullableSimpleString(final SimpleString val) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void writeNullableString(final String val) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void writeSimpleString(final SimpleString val) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void writeString(final String val) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void writeUTF(final String utf) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
@@ -1094,180 +665,6 @@ public class LargeMessageControllerImpl implements LargeMessageController {
    @Override
    public ActiveMQBuffer slice(final int index, final int length) {
       throw new UnsupportedOperationException();
-   }
-
-   /**
-    * @param output
-    * @param packet
-    * @throws ActiveMQException
-    */
-   private void sendPacketToOutput(final OutputStream output, final LargeData packet) throws ActiveMQException {
-      try {
-         output.write(packet.getChunk());
-         if (!packet.isContinues()) {
-            streamEnded = true;
-            output.close();
-         }
-      } catch (IOException e) {
-         throw ActiveMQClientMessageBundle.BUNDLE.errorWritingLargeMessage(e);
-      }
-   }
-
-   private void popPacket() {
-      try {
-
-         if (streamEnded) {
-            // no more packets, we are over the last one already
-            throw new IndexOutOfBoundsException();
-         }
-
-         int sizeToAdd = currentPacket != null ? currentPacket.chunk.length : 1;
-         currentPacket = largeMessageData.poll(readTimeout, TimeUnit.MILLISECONDS);
-         if (currentPacket == null) {
-            throw new IndexOutOfBoundsException();
-         }
-
-         if (currentPacket.chunk == null) { // Empty packet as a signal to interruption
-            currentPacket = null;
-            streamEnded = true;
-            throw new IndexOutOfBoundsException();
-         }
-
-         consumerInternal.flowControl(currentPacket.getFlowControlSize(), !currentPacket.isContinues());
-
-         packetPosition += sizeToAdd;
-
-         packetLastPosition = packetPosition + currentPacket.getChunk().length;
-      } catch (IndexOutOfBoundsException e) {
-         throw e;
-      } catch (Exception e) {
-         throw new RuntimeException(e);
-      }
-   }
-
-   private void checkForPacket(final long index) {
-      if (outStream != null) {
-         throw new IllegalAccessError("Can't read the messageBody after setting outputStream");
-      }
-
-      if (index >= totalSize) {
-         throw new IndexOutOfBoundsException();
-      }
-
-      if (streamClosed) {
-         throw new IllegalAccessError("The consumer associated with this large message was closed before the body was read");
-      }
-
-      if (fileCache == null) {
-         if (index < lastIndex) {
-            throw new IllegalAccessError("LargeMessage have read-only and one-way buffers");
-         }
-         lastIndex = index;
-      }
-
-      while (index >= packetLastPosition && !streamEnded) {
-         popPacket();
-      }
-   }
-
-   private final class FileCache {
-
-      private FileCache(final File cachedFile) {
-         this.cachedFile = cachedFile;
-      }
-
-      ByteBuffer readCache;
-
-      long readCachePositionStart = Integer.MAX_VALUE;
-
-      long readCachePositionEnd = -1;
-
-      private final File cachedFile;
-
-      private volatile FileChannel cachedChannel;
-
-      private synchronized void readCache(final long position) {
-
-         try {
-            if (position < readCachePositionStart || position > readCachePositionEnd) {
-
-               final FileChannel cachedChannel = checkOpen();
-
-               if (position > cachedChannel.size()) {
-                  throw new ArrayIndexOutOfBoundsException("position > " + cachedChannel.size());
-               }
-
-               readCachePositionStart = position / bufferSize * bufferSize;
-
-               cachedChannel.position(readCachePositionStart);
-
-               if (readCache == null) {
-                  readCache = ByteBuffer.allocate(bufferSize);
-               }
-
-               readCache.clear();
-
-               readCachePositionEnd = readCachePositionStart + cachedChannel.read(readCache) - 1;
-            }
-         } catch (Exception e) {
-            ActiveMQClientLogger.LOGGER.errorReadingCache(e);
-            throw new RuntimeException(e.getMessage(), e);
-         } finally {
-            close();
-         }
-      }
-
-      public synchronized byte getByteFromCache(final long position) {
-         readCache(position);
-
-         return readCache.get((int) (position - readCachePositionStart));
-
-      }
-
-      public void cachePackage(final byte[] body) throws Exception {
-         final FileChannel cachedChannel = checkOpen();
-
-         cachedChannel.position(cachedChannel.size());
-         cachedChannel.write(ByteBuffer.wrap(body));
-
-         close();
-      }
-
-      /**
-       * @throws FileNotFoundException
-       */
-      private FileChannel checkOpen() throws IOException {
-         FileChannel channel = cachedChannel;
-         if (cachedFile != null || !channel.isOpen()) {
-            channel = FileChannel.open(cachedFile.toPath(), StandardOpenOption.CREATE, StandardOpenOption.READ, StandardOpenOption.WRITE);
-            cachedChannel = channel;
-         }
-         return channel;
-      }
-
-      public void close() {
-         FileChannel cachedChannel = this.cachedChannel;
-         if (cachedChannel != null && cachedChannel.isOpen()) {
-            this.cachedChannel = null;
-            try {
-               cachedChannel.close();
-            } catch (Exception e) {
-               ActiveMQClientLogger.LOGGER.errorClosingCache(e);
-            }
-         }
-      }
-
-      @Override
-      protected void finalize() {
-         close();
-         if (cachedFile != null && cachedFile.exists()) {
-            try {
-               cachedFile.delete();
-            } catch (Exception e) {
-               ActiveMQClientLogger.LOGGER.errorFinalisingCache(e);
-            }
-         }
-      }
    }
 
    /**
@@ -1301,72 +698,42 @@ public class LargeMessageControllerImpl implements LargeMessageController {
 
    @Override
    public ActiveMQBuffer copy(final int index, final int length) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public ActiveMQBuffer duplicate() {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public ActiveMQBuffer readSlice(final int length) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void setChar(final int index, final char value) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void setDouble(final int index, final double value) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void setFloat(final int index, final float value) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public ActiveMQBuffer slice() {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
    @Override
    public void writeBytes(final ActiveMQBuffer src, final int srcIndex, final int length) {
-      throw new IllegalAccessError(LargeMessageControllerImpl.READ_ONLY_ERROR_MESSAGE);
+      throw new IllegalAccessError(READ_ONLY_ERROR_MESSAGE);
    }
 
-   private static class LargeData {
-
-      final byte[] chunk;
-      final int flowControlSize;
-      final boolean continues;
-
-      private LargeData() {
-         continues = false;
-         flowControlSize = 0;
-         chunk = null;
-      }
-
-      private LargeData(byte[] chunk, int flowControlSize, boolean continues) {
-         this.chunk = chunk;
-         this.flowControlSize = flowControlSize;
-         this.continues = continues;
-      }
-
-      public byte[] getChunk() {
-         return chunk;
-      }
-
-      public int getFlowControlSize() {
-         return flowControlSize;
-      }
-
-      public boolean isContinues() {
-         return continues;
-      }
-   }
 }
